Refactor: fix codes related vuex or firestore data
