+++ .vscode/settings.json
@@ -35,6 +35,7 @@
   "cSpell.words": [
     "Ouchi",
     "firestore",
-    "stylelint"
+    "stylelint",
+    "vuexfire"
   ]
 }
